More generally usable NFS permissions setup
Mysql load would fail with:

chown: changing ownership of '/var/lib/mysql': Operation not permitted

Changing the permisssions according to

https://docs.openshift.com/enterprise/3.0/admin_guide/persistent_storage_nfs.html

didn't fully help either -- this scenario assumes that the application
will be ok with simply reading/writing its data and not attempt
`chown`ing it. For kolla we actually need no_root_squash to map the
uid/gid of root to the NFS host, to allow doing chown.