Working MariaDB single-server
Using a bootstrap job to bootstrap the DB files. The resources can be
created in a single kubectl command but it seems that the statefulset
has to go first, otherwise the bootstrap job is stuck in creating
state (probably due to missing the PVC).

So:

    kubectl create -f services/mysql/statefulset.yaml -f services/mysql/bootstrap-job.yml

(assuming you've already created the service and configmap).